+++ python/example/remote_soem.py
@@ -14,10 +14,10 @@ Copyright (c) 2023 Shun Suzuki. All rights reserved.
 
 import asyncio
 
-from samples import runner
-
 from pyautd3 import AUTD3, Controller
 from pyautd3.link.soem import RemoteSOEM
+
+from .samples import runner
 
 
 async def main() -> None:
@@ -27,7 +27,7 @@ async def main() -> None:
         .open_with_async(
             RemoteSOEM.builder("127.0.0.1:8080"),
         )
-    ) as autd:
+    ) as autd:  # type: Controller
         await runner.run(autd)