+++ python/tests/conftest.py
@@ -16,28 +16,16 @@ import pytest
 
 
 def pytest_addoption(parser):
-    parser.addoption("--test_simulator", action="store_true", default=False, help="run simulator tests")
-    parser.addoption("--test_twincat", action="store_true", default=False, help="run twincat tests")
-    parser.addoption("--test_remote_twincat", action="store_true", default=False, help="run remote twincat tests")
-    parser.addoption("--test_soem", action="store_true", default=False, help="run soem tests")
-    parser.addoption("--test_remote_soem", action="store_true", default=False, help="run remote soem tests")
+    parser.addoption("--soem", action="store_true", default=False, help="run soem tests")
 
 
 def pytest_configure(config):
-    config.addinivalue_line("markers", "simulator: mark test as simulator test")
-    config.addinivalue_line("markers", "twincat: mark test as twincat test")
-    config.addinivalue_line("markers", "remote_twincat: mark test as remote twincat test")
-    config.addinivalue_line("markers", "soem: mark test as soem test")
-    config.addinivalue_line("markers", "remote_soem: mark test as remote soem test")
+    config.addinivalue_line("markers", "soem: soem tests")
 
 
 def pytest_collection_modifyitems(session, config, items):
     option_lists = [
-        ("--test_simulator", "simulator"),
-        ("--test_twincat", "twincat"),
-        ("--test_remote_twincat", "remote_twincat"),
-        ("--test_soem", "soem"),
-        ("--test_remote_soem", "remote_soem"),
+        ("--soem", "soem"),
     ]
     for option, marker in option_lists:
         if config.getoption(option):