+++ python/tests/gain/test_bessel.py
@@ -22,8 +22,13 @@ from tests.test_autd import create_controller
 async def test_bessel():
     autd = await create_controller()
 
-    assert await autd.send_async(Bessel(autd.geometry.center, [0, 0, 1], np.pi / 4).with_intensity(0x80))
+    assert await autd.send_async(Bessel(autd.geometry.center, [0, 0, 1], np.pi / 4))
+    for dev in autd.geometry:
+        intensities, phases = autd.link.intensities_and_phases(dev.idx, 0)
+        assert np.all(intensities == 0xFF)
+        assert not np.all(phases == 0)
 
+    assert await autd.send_async(Bessel(autd.geometry.center, [0, 0, 1], np.pi / 4).with_intensity(0x80))
     for dev in autd.geometry:
         intensities, phases = autd.link.intensities_and_phases(dev.idx, 0)
         assert np.all(intensities == 0x80)